Rename package: stores -> store
