+++ konashi-android-sdk/src/androidTest/java/com/uxxu/konashi/lib/store/PioStoreTest.java
@@ -1,4 +1,4 @@
-package com.uxxu.konashi.lib.stores;
+package com.uxxu.konashi.lib.store;
 
 import android.support.test.runner.AndroidJUnit4;
