+++ konashi-android-sdk/src/main/java/com/uxxu/konashi/lib/KonashiManager.java
@@ -34,11 +34,11 @@ import com.uxxu.konashi.lib.filter.AioAnalogReadFilter;
 import com.uxxu.konashi.lib.filter.BatteryLevelReadFilter;
 import com.uxxu.konashi.lib.filter.I2cReadFilter;
 import com.uxxu.konashi.lib.listeners.KonashiBaseListener;
-import com.uxxu.konashi.lib.stores.AioStore;
-import com.uxxu.konashi.lib.stores.I2cStore;
-import com.uxxu.konashi.lib.stores.PioStore;
-import com.uxxu.konashi.lib.stores.PwmStore;
-import com.uxxu.konashi.lib.stores.UartStore;
+import com.uxxu.konashi.lib.store.AioStore;
+import com.uxxu.konashi.lib.store.I2cStore;
+import com.uxxu.konashi.lib.store.PioStore;
+import com.uxxu.konashi.lib.store.PwmStore;
+import com.uxxu.konashi.lib.store.UartStore;
 
 import org.jdeferred.DoneCallback;
 import org.jdeferred.DonePipe;
