+++ konashi-android-sdk/src/main/java/com/uxxu/konashi/lib/dispatcher/PwmStoreUpdater.java
@@ -1,7 +1,7 @@
 package com.uxxu.konashi.lib.dispatcher;
 
 import com.uxxu.konashi.lib.KonashiUUID;
-import com.uxxu.konashi.lib.stores.PwmStore;
+import com.uxxu.konashi.lib.store.PwmStore;
 import com.uxxu.konashi.lib.util.PwmUtils;
 
 import java.util.UUID;
