+++ konashi-android-sdk/src/main/java/com/uxxu/konashi/lib/store/AioStore.java
@@ -1,4 +1,4 @@
-package com.uxxu.konashi.lib.stores;
+package com.uxxu.konashi.lib.store;
 
 import android.util.SparseArray;
