+++ konashi-android-sdk/src/main/java/com/uxxu/konashi/lib/store/UartStore.java
@@ -1,4 +1,4 @@
-package com.uxxu.konashi.lib.stores;
+package com.uxxu.konashi.lib.store;
 
 import com.uxxu.konashi.lib.Konashi;
 import com.uxxu.konashi.lib.dispatcher.CharacteristicDispatcher;
